revert trainOp in 2.2.0
