+++ python/orca/src/bigdl/orca/learn/pytorch/torch_runner.py
@@ -196,11 +196,11 @@ class TorchRunner(BaseRunner):
 
     def setup_ddp_components(self):
         from torch.nn.parallel import DistributedDataParallel
-        self.training_models = [
+        training_models = [
             DistributedDataParallel(model)
             for model in self.models
         ]
-        self.setup_operator(self.training_models)
+        self.setup_operator(training_models)
 
     def setup_operator(self, training_models):
         """Create the training operator."""
@@ -320,15 +320,15 @@ class TorchRunner(BaseRunner):
         })
         with self.timers.record("train_epoch"):
             data_loader = iter(data_loader)
-            train_stats = self._train_epoch(data_loader, info, callbacks)
+            train_stats = self.training_operator.train_epoch(data_loader, info, callbacks)
 
         if val_loader:
             with self.timers.record("validation"):
                 info = info or {}
-                validation_results = self._validate(val_loader,
-                                                    info=info,
-                                                    metrics=self.metrics,
-                                                    num_steps=val_steps)
+                validation_results = self.training_operator.validate(val_loader,
+                                                                     info=info,
+                                                                     metrics=self.metrics,
+                                                                     num_steps=val_steps)
                 # add prefix of "val_" for validation_stats
                 validation_stats = {}
                 for name, value in validation_results.items():
@@ -347,172 +347,6 @@ class TorchRunner(BaseRunner):
             stats.update(profile=self.timers.stats())
 
         return stats
-
-    def _train_epoch(self, iterator, info, callbacks=None):
-        """Runs one standard training pass over the training dataloader.
-
-        By default, this method will iterate over the given iterator and
-        call ``self.train_batch`` over each batch. If ``scheduler_step_freq``
-        is set, this default method will also step the scheduler accordingly.
-
-        You do not need to call ``train_batch`` in this method if you plan
-        to implement a custom optimization/training routine here.
-
-        You may find ``ray.util.sgd.utils.AverageMeterCollection`` useful
-        when overriding this method. See example below:
-
-        .. code-block:: python
-
-            def train_epoch(self, ...):
-                meter_collection = AverageMeterCollection()
-                self.model.train()
-                for batch in iterator:
-                    # do some processing
-                    metrics = {"metric_1": 1, "metric_2": 3} # dict of metrics
-
-                    # This keeps track of all metrics across multiple batches
-                    meter_collection.update(metrics, n=len(batch))
-
-                # Returns stats of the meters.
-                stats = meter_collection.summary()
-                return stats
-
-
-        Args:
-            iterator (iter): Iterator over the training data for the entire
-                epoch. This iterator is expected to be entirely consumed.
-            info (dict): Dictionary for information to be used for custom
-                training operations.
-
-        Returns:
-            A dict of metrics from training.
-        """
-        if self.use_tqdm and self.rank == 0:
-            desc = ""
-            if info is not None and "epoch_idx" in info:
-                if "num_epochs" in info:
-                    desc = "{}/{}e".format(info["epoch_idx"] + 1,
-                                           info["num_epochs"])
-                else:
-                    desc = "{}e".format(info["epoch_idx"] + 1)
-            _progress_bar = tqdm(
-                total=len(iterator),
-                desc=desc,
-                unit="batch",
-                leave=False)
-        else:
-            _progress_bar = None
-
-        metric_meters = AverageMeterCollection()
-
-        # TODO: Discuss the situation when there are multiple components,
-        #       It is best for the user to write this part of the logic in a hook func.
-        self.training_model.train()
-        # self.training_models may not be DDP if horovod.
-        from torch.nn.parallel import DistributedDataParallel as DDP
-        if isinstance(self.training_model, DDP):
-            with self.training_model.join():
-                self._train_loop(iterator, info, _progress_bar, metric_meters, callbacks)
-        else:
-            self._train_loop(iterator, info, _progress_bar, metric_meters, callbacks)
-
-        if self.scheduler and info.get(SCHEDULER_STEP) == SCHEDULER_STEP_EPOCH:
-            self.scheduler.step()
-
-        return metric_meters.summary(sync_stats=self.sync_stats,
-                                     dist_backend=self.dist_backend)
-
-    def _train_loop(self, iterator, info, _progress_bar, metric_meters, callbacks):
-        for batch_idx, batch in enumerate(iterator):
-            batch_info = {
-                "batch_idx": batch_idx,
-                "global_step": self.global_step
-            }
-            batch_info.update(info)
-            if callbacks is not None:
-                for callback in callbacks:
-                    callback.on_batch_begin(batch_idx)
-            metrics = self._train_batch(batch, batch_info=batch_info)
-            if self.use_tqdm and self.rank == 0:
-                _progress_bar.n = batch_idx + 1
-                postfix = {}
-                if "train_loss" in metrics:
-                    postfix.update(loss=metrics["train_loss"])
-                _progress_bar.set_postfix(postfix)
-
-            if self.scheduler and batch_info.get(
-                    SCHEDULER_STEP) == SCHEDULER_STEP_BATCH:
-                self.scheduler.step()
-
-            metric_meters.update(metrics, n=metrics.pop(NUM_SAMPLES, 1))
-            self.global_step += 1
-            if callbacks is not None:
-                for callback in callbacks:
-                    callback.on_batch_end(batch_idx, logs=metrics)
-
-    def _train_batch(self, batch, batch_info=None):
-        """Computes loss and updates the model over one batch.
-
-        This method is responsible for computing the loss and gradient and
-        updating the model.
-
-        By default, this method implementation assumes that batches
-        are in (\*features, labels) format. So we also support multiple inputs
-        model. If using amp/fp16 training, it will also scale the loss
-        automatically.
-
-        You can provide custom loss metrics and training operations if you
-        override this method. If overriding this method, you can access model,
-        optimizer, criterion via ``self.model``, ``self.optimizer``,
-        and ``self.criterion``.
-
-        You do not need to override this method if you plan to
-        override ``train_epoch``.
-
-        Args:
-            batch: One item of the validation iterator.
-            batch_info (dict): Information dict passed in from ``train_epoch``.
-
-        Returns:
-            A dictionary of metrics.
-                By default, this dictionary contains "loss" and "num_samples".
-                "num_samples" corresponds to number of datapoints in the batch.
-                However, you can provide any number of other values.
-                Consider returning "num_samples" in the metrics because
-                by default, ``train_epoch`` uses "num_samples" to
-                calculate averages.
-
-        """
-        # unpack features into list to support multiple inputs model
-        features, target = batch
-
-        # Compute output.
-        with self.timers.record("fwd"):
-            if torch.is_tensor(features):
-                output = self.training_model(features)
-            elif isinstance(features, (tuple, list)):
-                output = self.training_model(*features)
-            else:
-                invalidInputError(False,
-                                  "Features should be tensor, list/tuple or dict, "
-                                  "but got {}".format(type(features)))
-
-            if isinstance(output, tuple) or isinstance(output, list):
-                # Then target is also assumed to be a tuple or list.
-                loss = self.criterion(*output, *target)
-            else:
-                loss = self.criterion(output, target)
-
-        # Compute gradients in a backward pass.
-        with self.timers.record("grad"):
-            self.optimizer.zero_grad()
-            loss.backward()
-
-        # Call step of optimizer to update model params.
-        with self.timers.record("apply"):
-            self.optimizer.step()
-
-        return {"train_loss": loss.item(), NUM_SAMPLES: get_batchsize(features)}
 
     def validate(self, data_creator, batch_size=32, num_steps=None, profile=False,
                  info=None, wrap_dataloader=None):
@@ -535,106 +369,13 @@ class TorchRunner(BaseRunner):
             loader = self.with_sampler(loader)
         loader = iter(loader)
         with self.timers.record("validation"):
-            validation_stats = self._validate(loader,
-                                              info=info,
-                                              metrics=self.metrics,
-                                              num_steps=num_steps)
+            validation_stats = self.training_operator.validate(loader,
+                                                               info=info,
+                                                               metrics=self.metrics,
+                                                               num_steps=num_steps)
         if profile:
             validation_stats.update(profile=self.timers.stats())
         return validation_stats
-
-    def _validate(self, val_iterator, info, metrics, num_steps=None):
-        """Runs one standard validation pass over the val_iterator.
-
-        This will call ``model.eval()`` and ``torch.no_grad`` when iterating
-        over the validation dataloader.
-
-        If overriding this method, you can access model, criterion via
-        ``self.model`` and ``self.criterion``. You also do not need to call
-        ``validate_batch`` if overriding this method.
-
-        Args:
-            val_iterator (iter): Iterable constructed from the
-                validation dataloader.
-            info: (dict): Dictionary for information to be used for custom
-                validation operations.
-
-        Returns:
-            A dict of metrics from the evaluation.
-                By default, returns "val_accuracy" and "val_loss"
-                which is computed by aggregating "loss" and "correct" values
-                from ``validate_batch`` and dividing it by the sum of
-                ``num_samples`` from all calls to ``self.validate_batch``.
-        """
-        # switch to evaluate mode
-        self.model.eval()
-        metrics = Metric.convert_metrics_dict(metrics, backend="pytorch")
-        losses = []
-        total_samples = 0
-        with torch.no_grad():
-            for batch_idx, batch in enumerate(val_iterator):
-                if num_steps and batch_idx == num_steps:
-                    break
-                batch_info = {"batch_idx": batch_idx}
-                batch_info.update(info)
-                output, target, loss = self.forward_batch(batch, batch_info)
-                num_samples = get_batchsize(target)
-                total_samples += num_samples
-                losses.append(loss.item() * num_samples)
-                for metric in metrics.values():
-                    metric(output, target)
-
-        result = {name: metric.compute() for name, metric in metrics.items()}
-
-        result["val_loss"] = sum(losses) / total_samples
-
-        result["num_samples"] = total_samples
-
-        return result
-
-    def forward_batch(self, batch, batch_info):
-        """Calculates the loss and accuracy over a given batch.
-
-        You can override this method to provide arbitrary metrics.
-
-        Same as ``train_batch``, this method implementation assumes that
-        batches are in (\*features, labels) format by default. So we also
-        support multiple inputs model.
-
-        Args:
-            batch: One item of the validation iterator.
-            batch_info (dict): Contains information per batch from
-                ``validate()``.
-
-        Returns:
-            A dict of metrics.
-                By default, returns "val_loss", "val_accuracy", and
-                "num_samples". When overriding, consider returning
-                "num_samples" in the metrics because
-                by default, ``validate`` uses "num_samples" to
-                calculate averages.
-        """
-        # unpack features into list to support multiple inputs model
-        features, target = batch
-
-        # compute output
-        with self.timers.record("eval_fwd"):
-            if torch.is_tensor(features):
-                output = self.model(features)
-            elif isinstance(features, (tuple, list)):
-                output = self.model(*features)
-            else:
-                invalidInputError(False,
-                                  "Features should be tensor, list/tuple or dict, "
-                                  "but got {}".format(type(features)))
-
-            if isinstance(output, tuple) or isinstance(output, list):
-                # Then target is also assumed to be a tuple or list.
-                loss = self.criterion(*output, *target)
-            else:
-                loss = self.criterion(output, target)
-
-        return output, target, loss
 
     def predict(self, partition, batch_size=32, profile=False):
         """Evaluates the model on the validation data set."""
@@ -650,7 +391,7 @@ class TorchRunner(BaseRunner):
 
         def predict_fn(shard):
             if isinstance(partition, IterableDataset):
-                y = self._predict(shard)
+                y = self.training_operator.predict(shard)
             else:
                 if isinstance(shard["x"], tuple) or isinstance(shard["x"], list):
                     tensors = [torch.from_numpy(arr) for arr in shard["x"]]
@@ -658,7 +399,7 @@ class TorchRunner(BaseRunner):
                     tensors = [torch.from_numpy(shard["x"])]
                 dataset = torch.utils.data.TensorDataset(*tensors)
                 data_loader = DataLoader(dataset, **params)
-                y = self._predict(iter(data_loader))
+                y = self.training_operator.predict(iter(data_loader))
             return {"prediction": y}
 
         with self.timers.record("predict"):
@@ -667,34 +408,6 @@ class TorchRunner(BaseRunner):
             else:
                 new_part = [predict_fn(shard) for shard in partition]
         return new_part
-
-    def _predict(self, pred_iterator):
-        # switch to evaluate mode
-        self.model.eval()
-        result = []
-        with torch.no_grad():
-            for batch in pred_iterator:
-                result.append(self.predict_batch(batch))
-
-        return np.concatenate(result, axis=0)
-
-    def predict_batch(self, batch):
-
-        if isinstance(batch, torch.Tensor):
-            batch = [batch]
-
-        # compute output
-        with self.timers.record("pred_fwd"):
-            output = self.model(*batch)
-
-            if len(output.size()) > 1:
-                # In case there is extra trailing dimensions.
-                for i in reversed(range(1, len(output.size()))):
-                    output = torch.squeeze(output, i)
-
-        # todo support multi-output model
-        np_output = output.detach().numpy()
-        return np_output
 
     def _toggle_profiling(self, profile=False):
         """Enables/Disables and resets timing profiles."""
@@ -807,26 +520,3 @@ class TorchRunner(BaseRunner):
             return self.optimizers
         else:
             return self.optimizers[0]
-
-    @property
-    def model(self):
-        """First or only model(s) created by the ``model_creator``."""
-        if self.models:
-            return self.models[0]
-
-    @property
-    def optimizer(self):
-        """First or only optimizer(s) created by the ``optimizer_creator``."""
-        return self.optimizers[0]
-
-    @property
-    def scheduler(self):
-        """First or only scheduler(s) created by the ``scheduler_creator``."""
-        if self.schedulers:
-            return self.schedulers[0]
-
-    @property
-    def training_model(self):
-        """First or only training_model(s) wrapped by the torchDDP."""
-        if self.training_models:
-            return self.training_models[0]
